+++ src/command_getid.cpp
@@ -60,7 +60,7 @@ void CommandGetId::parse_id(const std::string& s) {
 }
 
 bool CommandGetId::setup(const std::vector<std::string>& arguments) {
-    po::options_description cmdline("Allowed options");
+    po::options_description cmdline("Available options");
     cmdline.add_options()
     ("id-file,i", po::value<std::string>(), "Read OSM IDs from given file")
     ;
@@ -85,6 +85,12 @@ bool CommandGetId::setup(const std::vector<std::string>& arguments) {
     po::variables_map vm;
     po::store(po::command_line_parser(arguments).options(desc).positional(positional).run(), vm);
     po::notify(vm);
+
+    if (vm.count("help")) {
+        std::cout << "Usage: osmium getid [OPTIONS] OSM-FILE ID...\n\n";
+        std::cout << cmdline << "\n";
+        exit(0);
+    }
 
     setup_common(vm);
     setup_input_file(vm);